requirements.txt
crewai
streamlit
openai
unstructured
tools
tenacity==8.3.0
streamlit-shadcn-ui
firebase-admin
langchain
langchain_groq
langchain_anthropic
anthropic
streamlit-extras
streamlit-cookies-manager
streamlit-custom-notification-box
cohere
borb
fpdf
python-dotenv
